+++ app/Http/Controllers/Admin/CourseController.php
@@ -8,6 +8,7 @@ use App\Http\Requests;
 use App\Http\Controllers\Controller;
 use App\Model\Course;
 use App\Model\Category;
+use App\Model\User;
 use Illuminate\Support\Facades\Storage;
 
 class CourseController extends Controller
@@ -19,8 +20,7 @@ class CourseController extends Controller
      */
     public function index()
     {
-        $datas = Course::orderBy('id', 'asc')->paginate(10);
-        // dd($data);
+        $datas = Course::orderBy('id', 'asc')->paginate(16);
         return view('admin.course.index',compact('datas',$datas));
     }
 
@@ -32,7 +32,8 @@ class CourseController extends Controller
     public function create()
     {
         $categorys = Category::select('id','name')->get();
-        return view('admin.course.create',['categorys'=>$categorys]);
+        $teachers = User::where('type',5)->get();
+        return view('admin.course.create',['categorys'=>$categorys,'teachers'=>$teachers]);
     }
 
     /**
@@ -41,11 +42,10 @@ class CourseController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(Requests\CreateCourseRequest $request)
     {
-        dd($request->all());
-        $filename = 'images/'.date('Y-m').'/'.date('YmdHis').mt_rand(10,99).'.png';
-        $res = Storage::put(
+        $filename = 'images/course/'.date('Y-m').'/'.date('YmdHis').mt_rand(10,99).'.png';
+        $res = Storage::disk('uploads')->put(
             $filename,
             file_get_contents($request->file('imgpath')->getRealPath())
         );
@@ -68,8 +68,8 @@ class CourseController extends Controller
             $data->is_recommend = 1;//已废弃（可更改为其他字段）
             $data->fee = 0.00;//暂不启用
             $data->status = 2;//默认关闭
-            $data->published_time = Carbon\Carbon::now();
-            $data->created_user = 1;
+            $data->published_time = date('Y-m-d H:i:s',time());
+            $data->created_user = \Auth::user()->id;
             $data->save();
             return redirect()->route('admin.course.index');
         }else{
@@ -98,8 +98,9 @@ class CourseController extends Controller
     public function edit($id)
     {
         $data = Course::find($id);
-        // dd($datas);
-        return view('admin.course.edit',compact('data',$data));
+        $categorys = Category::select('id','name')->get();
+        $teachers = User::where('type',5)->get();
+        return view('admin.course.edit',['data'=>$data,'teachers'=>$teachers,'categorys'=>$categorys]);
     }
 
     /**
@@ -109,18 +110,25 @@ class CourseController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(Requests\CreateCourseRequest $request, $id)
     {
         $course = Course::find($id);
-        // dd($course);
-        // dd($request->all());
+        if(!empty($request->imgpath)){
+            $filename = 'images/'.date('Y-m').'/'.date('YmdHis').mt_rand(10,99).'.png';
+            $res = Storage::disk('uploads')->put(
+                $filename,
+                file_get_contents($request->file('imgpath')->getRealPath())
+            );
+            if($res){
+                Storage::disk('uploads')->delete($course->imgpath);
+            }
+            $course->imgpath = $filename;
+        }
         $course->title = $request->title;
         $course->code = $request->code;
         $course->teacher_id = $request->teacher_id;
-        $course->is_recommend = $request->is_recommend;
         $course->difficulty = $request->difficulty;
-        $course->categorys = $request->categorys;
-        $course->imgpath = $request->imgpath;
+        $course->category_id = $request->category_id;
         $course->display_order = $request->display_order;
         $course->introduction = $request->introduction;
         $course->description = $request->description;
@@ -141,10 +149,11 @@ class CourseController extends Controller
 
     public function get_course_delete($id)
     {
-        $course = Course::find($id);
-
-        $course->delete();
-
+        $course = Course::findOrFail($id);
+        $res = Storage::disk('uploads')->delete($course->imgpath);
+        if($res){
+            $course->delete();
+        }
         return redirect()->route('admin.course.index');
     }
